refactor: reorganize option_base_*.go into option_bool/int/string/time/slice.go, merge OptionService struct into option.go
