+++ internal/core/option_base_activity.go
@@ -1,39 +1,1 @@
 package core
-
-import (
-	"fmt"
-)
-
-// WithActivityTypeIDs returns an option to set multiple `activityTypeId[]` parameters.
-func (s *OptionService) WithActivityTypeIDs(typeIDs []int) RequestOption {
-	return &APIParamOption{
-		Type: ParamActivityTypeIDs,
-		CheckFunc: func() error {
-			for _, id := range typeIDs {
-				if err := validateActivityID(id, "activityTypeIds"); err != nil {
-					return err
-				}
-			}
-			return nil
-		},
-		SetFunc: addIntFunc(ParamActivityTypeIDs, typeIDs),
-	}
-}
-
-// WithMaxID returns an option to set the `maxId` parameter.
-func (s *OptionService) WithMaxID(id int) RequestOption {
-	return intRangeOption(ParamMaxID, id, 1, MaxActivityTypeID)
-}
-
-// WithMinID returns an option to set the `minId` parameter.
-func (s *OptionService) WithMinID(id int) RequestOption {
-	return intRangeOption(ParamMinID, id, 1, MaxActivityTypeID)
-}
-
-// validateActivityID ensures that the given activity ID is within the valid range [1, 26].
-func validateActivityID(id int, key string) error {
-	if id < 1 || id > MaxActivityTypeID {
-		return NewValidationError(fmt.Sprintf("invalid %s: must be between 1 and %d", key, MaxActivityTypeID))
-	}
-	return nil
-}
