+++ internal/core/option_base_user.go
@@ -1,40 +1,1 @@
 package core
-
-import (
-	"github.com/nattokin/go-backlog/internal/model"
-)
-
-// WithMailAddress returns a option that sets the `mailAddress` field.
-func (s *OptionService) WithMailAddress(mailAddress string) RequestOption {
-	// ToDo: validate mailAddress (Note: The validation remains as simple not-empty check)
-	return nonEmptyStringOption(ParamMailAddress, mailAddress)
-}
-
-// WithPassword returns a option that sets the `password` field.
-func (s *OptionService) WithPassword(password string) RequestOption {
-	return &APIParamOption{
-		Type: ParamPassword,
-		CheckFunc: func() error {
-			if len(password) < 8 {
-				return NewValidationError("password must be at least 8 characters long")
-			}
-			return nil
-		},
-		SetFunc: setStringFunc(ParamPassword, password),
-	}
-}
-
-// WithRoleType returns a option that sets the `roleType` field.
-func (s *OptionService) WithRoleType(roleType model.Role) RequestOption {
-	return intRangeOption(ParamRoleType, int(roleType), 1, 6)
-}
-
-// WithSendMail returns a option to specify whether to send an invitation email.
-func (s *OptionService) WithSendMail(enabled bool) RequestOption {
-	return boolOption(ParamSendMail, enabled)
-}
-
-// WithUserID returns a option to set the user's ID.
-func (s *OptionService) WithUserID(id int) RequestOption {
-	return positiveIntOption(ParamUserID, id)
-}
